Some optimizations to block importer
