+++ node/consensus/Cargo.toml
@@ -11,7 +11,7 @@ client = { path = "../../node/client" }
 primitives = { path = "../../core/primitives" }
 transaction = { path = "../../core/transaction" }
 substrate-network-libp2p = { git = "https://github.com/nearprotocol/substrate", rev = "2db35a0c6e87fc4450a77da548ef6e2db8b0d057" }
-beacon-chain-handler = { path = "../beacon-chain-handler" }
+coroutines = { path = "../coroutines" }
 txflow = { path = "../../core/txflow" }
 typed-arena = "1.4.1"
 futures = "0.1.25"
