+++ node/consensus/src/lib.rs
@@ -12,7 +12,7 @@ extern crate typed_arena;
 extern crate primitives;
 extern crate network;
 extern crate beacon;
-extern crate beacon_chain_handler;
+extern crate coroutines;
 extern crate substrate_network_libp2p;
 extern crate txflow;
 pub mod adapters;
