+++ node/coroutines/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "beacon-chain-handler"
+name = "coroutines"
 version = "0.1.0"
 edition = "2018"
 
@@ -12,9 +12,6 @@ chain = { path = "../../core/chain" }
 client = { path = "../../node/client" }
 primitives = { path = "../../core/primitives" }
 transaction = { path = "../../core/transaction" }
-node-runtime = { path = "../runtime" }
-storage = { path = "../../core/storage" }
-parking_lot = "0.6"
 futures = "0.1.25"
 tokio = "0.1.11"
 txflow = { path = "../../core/txflow" }
